Project_Auth/requirements.txt
fastapi==0.115.4
python-jose==3.3.0
pydantic
datetime
passlib==1.7.4
uvicorn==0.32.0
sqlmodel==0.0.22
sqlalchemy
pydantic[email]
typing
mysqlclient==2.2.6
python-dotenv==1.0.1
bcrypt